image_classification/requirements.txt
executorch
torchvision
Pillow
opencv-python
numpy
python-dotenv
torch
nltk
